review: isolate metrics display into loops_run_metrics_viewer
Extracts the cli-table layout, NDJSON emission, and live-tail loop into
loops_run_metrics_viewer.py, mirroring the existing loops_checkpoint_viewer
pattern. The command body in loops_commands.py now reads as wiring +
dispatch (parse args → resolve trainer deployment → build fetch closure →
delegate to render_metrics_snapshot / tail_metrics_table / emit_json_snapshots).
