+++ truss/cli/loops_commands.py
@@ -1,12 +1,8 @@
 import json
 import re
-import sys
-import time
-from contextlib import contextmanager
 from datetime import datetime, timedelta
-from typing import Any, Callable, Dict, Iterator, List, Optional, TextIO, Tuple, cast
+from typing import Any, Dict, List, Optional, Tuple, cast
 
-import rich.live
 import rich.table
 import rich_click as click
 import yaml
@@ -20,6 +16,12 @@ from truss.cli.logs.model_log_watcher import ModelDeploymentLogWatcher
 from truss.cli.loops_checkpoint_viewer import (
     resolve_most_recent_run_for_base_model,
     view_loops_checkpoint_list,
+)
+from truss.cli.loops_run_metrics_viewer import (
+    DEFAULT_METRICS_REFRESH_SECONDS,
+    emit_json_snapshots,
+    render_metrics_snapshot,
+    tail_metrics_table,
 )
 from truss.cli.train import checkpoint_viewer as checkpoint_mod
 from truss.cli.utils import common
@@ -221,9 +223,6 @@ def view_loops_runs(
     _render_loops_runs(runs)
 
 
-_DEFAULT_METRICS_REFRESH_SECONDS = 30
-_SPARKLINE_BLOCKS = "▁▂▃▄▅▆▇█"
-_SPARKLINE_WIDTH = 24
 _DURATION_RE = re.compile(r"^\s*(\d+)\s*([smhd])\s*$", re.IGNORECASE)
 _DURATION_UNIT_SECONDS = {"s": 1, "m": 60, "h": 3600, "d": 86400}
 # Deployments in these states have no live pods producing metrics, so they
@@ -245,7 +244,7 @@ _INACTIVE_DEPLOYMENT_STATUSES = frozenset({"FAILED", "STOPPED", "SCALED_TO_ZERO"
 @click.option(
     "--refresh-rate-seconds",
     type=int,
-    default=_DEFAULT_METRICS_REFRESH_SECONDS,
+    default=DEFAULT_METRICS_REFRESH_SECONDS,
     show_default=True,
     help="Seconds between refreshes when --tail is set.",
 )
@@ -379,7 +378,7 @@ def view_loops_run_metrics(
         }
 
     if is_json:
-        _emit_metrics_json(
+        emit_json_snapshots(
             fetch_snapshot,
             tail=tail,
             refresh_rate_seconds=refresh_rate_seconds,
@@ -393,19 +392,12 @@ def view_loops_run_metrics(
             f"Fetching metrics for Loops run [cyan]{run_id}[/cyan]...", spinner="dots"
         ):
             snapshot = fetch_snapshot()
-        console.print(_build_metrics_table(snapshot))
+        render_metrics_snapshot(snapshot)
         return
 
-    with rich.live.Live(auto_refresh=False, console=console, screen=False) as live:
-        try:
-            while True:
-                live.update(_build_metrics_table(fetch_snapshot()), refresh=True)
-                time.sleep(refresh_rate_seconds)
-        except KeyboardInterrupt:
-            live.stop()
-            console.print(
-                f"\nStopped watching metrics for run {run_id}.", style="yellow"
-            )
+    tail_metrics_table(
+        fetch_snapshot, refresh_rate_seconds=refresh_rate_seconds, run_id=run_id
+    )
 
 
 def _resolve_trainer_deployment_for_run(
@@ -449,126 +441,6 @@ def _iso_to_epoch_millis(iso_timestamp: str) -> int:
     return int(
         datetime.fromisoformat(iso_timestamp.replace("Z", "+00:00")).timestamp() * 1000
     )
-
-
-def _latest_value(points: List[Dict[str, Any]]) -> Optional[float]:
-    if not points:
-        return None
-    return points[-1].get("value")
-
-
-def _sparkline(points: List[Dict[str, Any]]) -> str:
-    values = [p["value"] for p in points if p.get("value") is not None]
-    if not values:
-        return ""
-    width = min(_SPARKLINE_WIDTH, len(values))
-    if len(values) > width:
-        bucket = len(values) / width
-        bucketed = []
-        for i in range(width):
-            lo_idx = int(i * bucket)
-            hi_idx = max(int((i + 1) * bucket), lo_idx + 1)
-            chunk = values[lo_idx:hi_idx]
-            bucketed.append(sum(chunk) / len(chunk))
-    else:
-        bucketed = values
-    lo, hi = min(bucketed), max(bucketed)
-    if hi - lo <= 0:
-        return _SPARKLINE_BLOCKS[len(_SPARKLINE_BLOCKS) // 2] * len(bucketed)
-    return "".join(
-        _SPARKLINE_BLOCKS[int((v - lo) / (hi - lo) * (len(_SPARKLINE_BLOCKS) - 1))]
-        for v in bucketed
-    )
-
-
-def _short_iso(iso: str) -> str:
-    try:
-        return (
-            datetime.fromisoformat(iso.replace("Z", "+00:00"))
-            .astimezone()
-            .strftime("%Y-%m-%d %H:%M:%S")
-        )
-    except ValueError:
-        return iso
-
-
-def _fmt_latest(value: Optional[float]) -> str:
-    if value is None:
-        return "—"
-    return f"{value:.2f}"
-
-
-def _build_metrics_table(snapshot: Dict[str, Any]) -> rich.table.Table:
-    window = snapshot["window"]
-    table = rich.table.Table(
-        show_header=True,
-        header_style="bold magenta",
-        title=(
-            f"Metrics for Loops run [cyan]{snapshot['run_id']}[/cyan]  "
-            f"({_short_iso(window['start'])} → {_short_iso(window['end'])})"
-        ),
-        box=rich.table.box.ROUNDED,
-        border_style="blue",
-    )
-    table.add_column("Component", style="green")
-    table.add_column("Deployment", style="cyan")
-    table.add_column("Request volume (req/s)", justify="right")
-    table.add_column("Concurrent requests", justify="right")
-    table.add_column("Trend", justify="left")
-
-    trainer = snapshot["trainer"]
-    table.add_row(
-        "Trainer",
-        snapshot["trainer_deployment_id"],
-        _fmt_latest(_latest_value(trainer["request_volume"])),
-        _fmt_latest(_latest_value(trainer["concurrent_requests"])),
-        _sparkline(trainer["request_volume"]),
-    )
-
-    sampler_id = snapshot["sampler_deployment_id"]
-    sampler = snapshot["sampler"]
-    if sampler_id:
-        table.add_row(
-            "Sampler",
-            sampler_id,
-            _fmt_latest(_latest_value(sampler["request_volume"])),
-            _fmt_latest(_latest_value(sampler["concurrent_requests"])),
-            _sparkline(sampler["request_volume"]),
-        )
-    else:
-        table.add_row("Sampler", "—", "—", "—", "")
-    return table
-
-
-@contextmanager
-def _open_metrics_sink(output_file: Optional[str]) -> Iterator[TextIO]:
-    if output_file is None or output_file == "-":
-        yield sys.stdout
-        return
-    with open(output_file, "w") as f:
-        yield f
-
-
-def _emit_metrics_json(
-    fetch_snapshot: Callable[[], Dict[str, Any]],
-    *,
-    tail: bool,
-    refresh_rate_seconds: int,
-    output_file: Optional[str],
-    run_id: str,
-) -> None:
-    with _open_metrics_sink(output_file) as sink:
-        if not tail:
-            sink.write(json.dumps(fetch_snapshot()) + "\n")
-            sink.flush()
-            return
-        try:
-            while True:
-                sink.write(json.dumps(fetch_snapshot()) + "\n")
-                sink.flush()
-                time.sleep(refresh_rate_seconds)
-        except KeyboardInterrupt:
-            print(f"Stopped streaming metrics for run {run_id}.", file=sys.stderr)
 
 
 @loops.group(name="samplers")
